i added minor customerRepository and added the customer profile page and i added minor changes to the customer class and i added the customer service and customer controller.

+++ waha/src/main/java/com/bookstore/waha/model/Customers.java
@@ -61,12 +61,12 @@ public class Customer {
 		this.password = password;
 	}
 
-	public Integer getCustomerID() {
+	public Long getCustomerID() {
 		return customerID;
 	}
 
 
-	public void setCustomerID(Integer customerID) {
+	public void setCustomerID(Long customerID) {
 		this.customerID = customerID;
 	}
